feat: 상세화면 타이틀 구현(코인 한글 이름 + 심볼) #32
