+++ Huhoe/Huhoe/Presentation/Scene/HuhoeMain/View/CoinListCell.swift
@@ -68,34 +68,3 @@ extension CoinListCell {
         }
     }
 }
-
-private extension NSMutableAttributedString {
-    func text(
-        _ string: String,
-        fontStyle: UIFont.TextStyle = .title3
-    ) -> NSMutableAttributedString {
-        let font = UIFont.preferredFont(forTextStyle: fontStyle).bold
-        let attributes: [NSAttributedString.Key: Any] = [.font: font]
-        
-        self.append(NSAttributedString(string: string, attributes: attributes))
-        
-        return self
-    }
-
-    func colorText(
-        _ string: String,
-        color: UIColor,
-        fontStyle: UIFont.TextStyle = .title3
-    ) -> NSMutableAttributedString {
-        let font = UIFont.preferredFont(forTextStyle: fontStyle).bold
-        let attributes:[NSAttributedString.Key : Any] = [
-            .font: font,
-            .foregroundColor: color
-        ]
-
-        self.append(NSAttributedString(string: string, attributes:attributes))
-        
-        return self
-    }
-}
-
